fix(cli): keep gateway-backed agent turns cold
