+++ src/commands/channels.ts
@@ -12,4 +12,4 @@ export { channelsRemoveCommand } from "./channels/remove.js";
 export type { ChannelsResolveOptions } from "./channels/resolve.js";
 export { channelsResolveCommand } from "./channels/resolve.js";
 export type { ChannelsStatusOptions } from "./channels/status.js";
-export { channelsStatusCommand, formatGatewayChannelsStatusLines } from "./channels/status.js";
+export { channelsStatusCommand } from "./channels/status.js";
